rent/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.0.1
django-ckeditor==6.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-tinymce==3.4.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.5
python-environ==0.4.54
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
